Add files via upload
This is the base file (updated as at Feb 21, 2020) for the DIY Excel Accounting System.

I created this base file mainly for demonstrating how to build an accounting system from scratch.
Tailor-made features, included but not limited to, foreign currency, multilingual interface, audit pack, tax pack, analytical pack, are missing.

Further videos will be provided to show how to add modules to the accounting system for your need.

Please note that the system is NOT MEANT TO BE FOR PRODUCTION.  
Users are solely responsible for any impact caused by using or sharing the files, and that caused by following my videos.

Feel free to build your own system.  Hope it helps.